pyproject.toml
[tool.poetry]
name = "workerfacing-api"
version = "0.1.0"
description = "Worker-facing API of DECODE OpenCloud."
authors = ["Arthur Jaques <arthur.jaques@hispeed.ch>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.10"
boto3 = "^1.35.41"
botocore = "^1.35.41"
deprecated = "^1.2.14"
python-dotenv = "^1.0.1"
dict-hash = "^1.3.4"
fastapi = "^0.115.2"
fastapi-cloudauth = "^0.4.3"
fastapi-utils = "^0.7.0"
httpx = "^0.27.2"
pydantic = "^2.9.2"
python-multipart = "^0.0.12"
requests = "^2.32.3"
gunicorn = "^23.0.0"
uvicorn = "^0.32.0"
typing-inspect = "^0.9.0"
psycopg2-binary = "^2.9.10"
sqlalchemy = "^2.0.36"

[tool.poetry.group.dev.dependencies]
boto3-stubs = {extras = ["essential"], version = "^1.35.54"}
botocore-stubs = "^1.35.41"
moto = "^5.0.17"
mypy = "^1.12.0"
pre-commit = "^4.0.1"
pytest = "^8.3.3"
ruff = "^0.6.9"
types-deprecated = "^1.2.9.20240311"
types-requests = "^2.32.0.20240914"
pytest-cov = "^5.0.0"
docker = "^7.1.0"
toml = "^0.10.2"
types-toml = "^0.10.8.20240310"
types-docker = "^7.1.0.20240827"
types-boto3 = "^1.0.2"
ipython = "^8.29.0"

[build-system]
requires = ["poetry-core=1.8.4"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
docker-build = "scripts.docker:build"
docker-serve = "scripts.docker:serve"
docker-cleanup = "scripts.docker:cleanup"
serve = "scripts.serve:main"

[tool.pytest.ini_options]
markers = [
    "aws: requires aws credentials"
]
addopts = "-m 'not aws'"

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.lint.per-file-ignores]
"tests/conftest.py" = ["E402"]
"workerfacing_api/main.py" = ["E402"]

[tool.mypy]
strict = true
